backend/pyproject.toml
[tool.poetry]
name = "finance-recon-ai-backend"
version = "0.1.0"
description = "Backend for Finance Recon AI"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
langchain = {version = "^1.2.0", allow-prereleases = true}
langchain-community = "^0.4.1"
langchain-core = "^1.2.5"
langchain-ollama = "^1.0.1"
langchain-openai = "^1.1.6"
langgraph = "^1.0.5"
ollama = "^0.6.1"
pandas = "^2.2.1"
pdfplumber = "^0.10.4"
pypdf = "^4.0.0"
aiofiles = "^23.2.0"
thefuzz = "^0.22.1"
faiss-cpu = "^1.8.0"
langchain-google-genai = "^4.1.2"
google-generativeai = "^0.8.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
